+++ src/bract/core/util.clj
@@ -31,26 +31,6 @@
     (apply println x more)))
 
 
-(defn apply-inducer
-  "Given a context and an inducer function `(fn [context]) -> context`, apply the inducer to the context returning
-  an updated context."
-  [context inducer]
-  (inducer context))
-
-
-(defn induce
-  "Given a seed context (optional, {} by default), a reducing function `(fn [context inducer-candidate]) -> context`
-  (optional, `bract.core.util/apply-inducer` by default) and a collection of inducer candidates, roll the seed context
-  through each inducer function successively, returning an updated context."
-  ([coll]
-    (induce {} apply-inducer coll))
-  ([seed coll]
-    (induce seed apply-inducer coll))
-  ([seed f coll]
-    (reduce (fn [context inducer-candidate] (f context inducer-candidate))
-      seed coll)))
-
-
 (defn shorten-name
   "Shorten a stringable name, e.g. foo.bar.baz.qux/fred to f.b.baz.qux/fred, leaving only the last two tokens intact."
   ([any-name]
@@ -76,3 +56,11 @@
      (alter-var-root var# (fn [old-val#] (or old-val# (do (vreset! old# old-val#) true))))
      (when-not @old#
        ~@body)))
+
+
+(defn as-vec
+  "Turn argument into a vector if it is a collection, else wrap the argument into a single-item vector."
+  [x]
+  (if (coll? x)
+    (vec x)
+    [x]))