+++ test/bract/core/config_test.clj
@@ -17,18 +17,21 @@
   (testing "happy cases"
     (let [good-context {:bract.core/verbose? false
                         :bract.core/config-files ["foo.edn" "bar.properties"]
+                        :bract.core/exit? :yes
                         :bract.core/cli-args ["-c" "run"]
                         :bract.core/config {"foo" "bar"}
                         :bract.core/launch? true}]
       (is (false? (config/ctx-verbose? good-context)))
       (is (= ["foo.edn" "bar.properties"] (config/ctx-config-files good-context)))
+      (is (= :yes (config/ctx-exit? good-context)))
       (is (= ["-c" "run"] (config/ctx-cli-args good-context)))
       (is (= {"foo" "bar"} (config/ctx-config good-context)))
       (is (= true (config/ctx-launch? good-context)))
       (is (false? (config/ctx-launch? {})) "missing/default value")))
   (testing "default values"
-    (is (true? (config/ctx-verbose? {})))
-    (is (= [] (config/ctx-config-files {}))))
+    (is (false? (config/ctx-verbose? {})))
+    (is (= [] (config/ctx-config-files {})))
+    (is (false? (config/ctx-exit? {}))))
   (testing "missing/bad context keys"
     (let [bad-context {:bract.core/config "foobar"
                         :bract.core/launch? 10}]
@@ -41,25 +44,17 @@
   (testing "happy tests"
     (doseq [good-config [{"bract.core.inducers" ["foo.bar.baz.qux/fred"
                                                  "mary.had.a.little/lamb"]
-                          "bract.core.context-hook" 'bract.core.config/apply-inducer-by-name
-                          "bract.core.config-hook"  'bract.core.config/apply-inducer-by-name
                           "bract.core.exports" ["foo"
                                                 "bar"]
-                          "bract.core.launcher" 'bract.core.config/apply-inducer-by-name}
+                          "bract.core.launcher" 'bract.core.inducer/apply-inducer}
                          {"bract.core.inducers" "[\"foo.bar.baz.qux/fred\"
                                                    \"mary.had.a.little/lamb\"]"
-                          "bract.core.context-hook" "bract.core.config/apply-inducer-by-name"
-                          "bract.core.config-hook"  "bract.core.config/apply-inducer-by-name"
                           "bract.core.exports"  "[\"foo\"
                                                    \"bar\"]"
-                          "bract.core.launcher" "bract.core.config/apply-inducer-by-name"}]]
+                          "bract.core.launcher" "bract.core.inducer/apply-inducer"}]]
       (is (= ["foo.bar.baz.qux/fred"
               "mary.had.a.little/lamb"]
             (config/cfg-inducers good-config)))
-      (is (some?
-            (config/cfg-context-hook good-config)))
-      (is (some?
-            (config/cfg-config-hook good-config)))
       (is (= ["foo" "bar"]
             (config/cfg-exports good-config)))
       (is (some?
@@ -72,10 +67,6 @@
                        "bract.core.launcher" false}]
       (is (thrown? IllegalArgumentException (config/cfg-inducers      bad-context)))
       (is (thrown? IllegalArgumentException (config/cfg-inducers      {})) "missing key")
-      (is (thrown? IllegalArgumentException (config/cfg-context-hook  bad-context)))
-      (is (thrown? IllegalArgumentException (config/cfg-context-hook  {})) "missing key")
-      (is (thrown? IllegalArgumentException (config/cfg-config-hook   bad-context)))
-      (is (thrown? IllegalArgumentException (config/cfg-config-hook   {})) "missing key")
       (is (thrown? IllegalArgumentException (config/cfg-exports       bad-context)))
       (is (thrown? IllegalArgumentException (config/cfg-exports       {})) "missing key")
       (is (thrown? IllegalArgumentException (config/cfg-launcher      bad-context)))